+++ src/PascalVocTools.jl
@@ -1,10 +1,11 @@
 module PascalVocTools
 
 
-export voc_colors, voc_classnumbers, voc_classnames,
-       voc_colormaps2classnumbers, voc_classnumbers2colormaps,
-       voc_rgb2classes,
-       voc_download
+export voc_colors,
+        voc_classnumber2classname, voc_classname2classnumber,
+        voc_colormap2classnumber, voc_classnumber2colormap,
+        voc_rgb2classes,
+        voc_download
 
 import Images: RGB, Gray, N0f8, channelview
 import Downloads