+++ test/conversions_tests.jl
@@ -1,14 +1,14 @@
-@test voc_classnumbers(10) == "cow"
-@test voc_classnumbers(30) == "void"  # 30 is not a valid class, should return void
+@test voc_classnumber2classname(10) == "cow"
+@test voc_classnumber2classname(30) == "void"  # 30 is not a valid class, should return void
 
-@test voc_classnames("cow") == 10
-@test voc_classnames("unknown") == 255  # "unknown" is not a valid class, should return 255 (void)
+@test voc_classname2classnumber("cow") == 10
+@test voc_classname2classnumber("unknown") == 255  # "unknown" is not a valid class, should return 255 (void)
 
-@test voc_colormaps2classnumbers((64, 128, 0)) == 10
-@test voc_colormaps2classnumbers((20, 30, 40)) == 255  # (20, 30, 40) is not a valid colormap, should return 255 (void)
+@test voc_colormap2classnumber((64, 128, 0)) == 10
+@test voc_colormap2classnumber((20, 30, 40)) == 255  # (20, 30, 40) is not a valid colormap, should return 255 (void)
 
-@test voc_classnumbers2colormaps(10) == (64, 128, 0)
-@test voc_classnumbers2colormaps(30) == (224, 224, 192)  # 30 is not a valid class, should return void color
+@test voc_classnumber2colormap(10) == (64, 128, 0)
+@test voc_classnumber2colormap(30) == (224, 224, 192)  # 30 is not a valid class, should return void color
 
 c = (0, 0, 0) ./ 255
 c = Images.RGB{N0f8}(c...)