+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v0.0.3
-* Change color representations from vectors to triplets [...] -> (...)
+* Change color representations from vectors to triplets [...] -> (...);
+* Added voc_colors colormap.
 
 ### v0.0.2
 * Void color corrected to [224, 224, 192].